third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'y' property of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.y.baseVal.value is -80
PASS foreignObjectElement.y.baseVal.value is 80
PASS successfullyParsed is true

TEST COMPLETE
